Moved ColorChooser to folder widgets/
