+++ example/lib/widgets/color_chooser.dart
